Added repeatability for Marking Bad Channels
Added ability for user to indicate whether they want to repeat the bad channel marking step.  Also made it so that user is notified of when data is too short and warned when >5% of channels are marked as bad.  Also fixed string bug in listing paradigms that occurred in Windows environments.