Add support for calculating two dipoles
Previously, the code only supported calculating one dipole. This commit adds support for calculating two dipoles by introducing a new function `calculate_two_dipoles`. If the number of dipoles is set to 2, the function will be called and it will calculate two dipoles using the `fitTwoDipoles` plugin with specific parameters. The availability of the plugin is also checked and logged accordingly.

This change allows for more flexibility in dipole calculations and improves the functionality of the code.