feat: Add option to compute channel number instead of component number
This commit adds a new option to the code that allows for computing the channel number instead of the component number. The `computeNumComponentsOrChannels` function has been modified to include this functionality. When the `use_components` flag is set to false, the code will now compute the channel number instead of the component number. This change provides more flexibility in analyzing EEG data.