feat: Add options for parallel processing and recalculation of fooof
This commit adds two new optional parameters to the `eeg_htpVisualizeIcProps` function: `parallel` and `recalculate_fooof`. The `parallel` parameter allows the user to specify whether or not to run certain calculations in parallel, while the `recalculate_fooof` parameter determines whether or not to recalculate fooof. These options provide more flexibility and control over the visualization of IC properties.